Silence rails engine warnings
MSP-9228

When I removed quietly around the eager load of Metasploit::Model and
MetasploitDataModels, I allowed the Engine not loading due to rails not
loading messages to be printed to stderr when msfconsole starts.
Instead of going back to quietly and potentially hiding new warnings,
adding silence_rails_engine_warnings, which will only not print the
current rails warnings.