Metasploit::Framework::Deprecation
MSP-9228

Metasploit::Framework::Deprecation.rename_methods can be used to
automatically declare methods that issue
Metasploit::Framework::Deprecation.warn and call the new method name.
Metasploit::Framework::Deprecation.warn wraps
ActiveSupport::Deprecation.warn to use a standard format for the message
and reports the deprecated method on the owner that declared the method
instead of the receiver of the method.