Enumerator for converting Mdm::Module::Class to Msf::Module
MSP-9228

A common pattern in the new code is needing to take an
ActiveRecord::Relation<Mdm::Module::Class> and attempting to instantiate
each class using framework.modules.create_from_module_class to get
metadata that is not yet in the module cache.
Metasploit::Framework::Module::Instance::Enumerator bundle this pattern
up and includes logging when an Mdm::Module::Class fails to create an
instance.