Use framework.pathnames.database_yaml for DatabaseConnection#database_yaml_pathname
MSP-9228

Add Metasploit::Framework::Framework::Pathnames#database_yaml, which
is used in place of the path calculation on
Metasploit::Framework::DatabaseConnection.  The
:metasploit_framework_framework_pathnames factory is enhanced to call
 #make after build and then to copy config/database.yml to
 #database_yaml, so the factory built pathnames can still access the
test database.  The 'Msf::Simple::Framework' shared context is
simplified to just use the :msf_simple_framework factory for #framework
since the factory is now creating the pathnames.