Optional tab completion for 'Rex::Ui::Text::DispatcherShell command'
MSP-11128

Not all commands have tab completion, such as when they take no
arguments, so have a boolean `:tab_completion` option that needs to be
set as `true` or `false`.  If `true`, then expect to respond to the
tab completion method; if `false`, then expect not to respond to the tab
completion method.